@@ -26,7 +26,7 @@ export class ConfigComponent extends BaseComponent implements OnInit, OnDestroy 
 
   selectedItem: NodeInstance;
 
-  @ViewChild("configTree", { static: false })
+  @ViewChild("configTree")
   configTree: ConfigTreeComponent;
 
   deletedConfig: number[] = [];
